# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PhishGuard API",
        "description": "API untuk deteksi dan pelaporan phishing",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://be-phisguard-production.up.railway.app/api",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles",
                "operationId": "89181b74a8aa4d580750ede6aa32f92a",
                "parameters": [
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Tips",
                                "Modus",
                                "Update Kasus"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create new educational article",
                "operationId": "ce06e2544b80bc76b184c771c0fb2e7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "kategori_artikel",
                                    "isi_artikel",
                                    "rangkuman"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Waspada Phishing Via Email"
                                    },
                                    "kategori_artikel": {
                                        "type": "string",
                                        "example": "Tips",
                                        "enum": [
                                            "Tips",
                                            "Modus",
                                            "Update Kasus"
                                        ]
                                    },
                                    "gambar": {
                                        "description": "Image file (JPEG/PNG)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "alt_text": {
                                        "type": "string",
                                        "example": "Illustrasi email phishing"
                                    },
                                    "isi_artikel": {
                                        "type": "string",
                                        "example": "Artikel lengkap tentang cara mendeteksi phishing..."
                                    },
                                    "rangkuman": {
                                        "type": "string",
                                        "example": "Ringkasan singkat artikel"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get article by ID",
                "operationId": "126f0079c2def3834cf2c4a2456d16f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article detail"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete article",
                "operationId": "ec28483335ebdc4fd11f883fcf6f4487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted successfully"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update educational article",
                "operationId": "98182cbd13556214028a5d193e7ebc5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Waspada Phishing Via Email"
                                    },
                                    "kategori_artikel": {
                                        "type": "string",
                                        "enum": [
                                            "Tips",
                                            "Modus",
                                            "Update Kasus"
                                        ]
                                    },
                                    "gambar": {
                                        "description": "Image file (JPEG/PNG)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "alt_text": {
                                        "type": "string"
                                    },
                                    "isi_artikel": {
                                        "type": "string"
                                    },
                                    "rangkuman": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article updated successfully"
                    },
                    "404": {
                        "description": "Article not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
...
    },
...
}